// mobile/types/character_types.ts
export type Attributes = {
    resistance: number;
    strength: number;
    agility: number;
    dexterity: number;
    vivacity: number;
    intellect: number;
    social: number;
    magic: number;
    [key: string]: number;
};
// ...
export type Religion = {
    god: string;
    devotion: number;
    [key: string]: any;
};

export type Skill = {
    name: string;
    category: string;
    pureValue: number;
    roleValue: number;
    [key: string]: any;
};

export type BasicCharaInfo = {
    name: string;
    age: number;
    race: string;
    gender: string;
    [key: string]: any;
};
// ...
export type CharacterOtherInfo = {
    languages: string[];
    experience: number;
    mana:       number;
    money:      number;
    [key: string]: any;
}

export type Priority = {
    role: string;
    attribute: string;
    skills: string;
    money: string;
}